Add StateManager and tests
